Move generated Dart client out of lib/ to fix language version conflict
