+++ pallas_app/analysis_options.yaml
@@ -11,7 +11,7 @@
 include: package:flutter_lints/flutter.yaml
 analyzer:
   exclude:
-    - lib/proxy/generated/**
+    - packages/openapi/**
 linter:
   # The lint rules applied to this project can be customized in the
   # section below to disable rules from the `package:flutter_lints/flutter.yaml`
