+++ pallas_app/lib/proxy/client/api_client.dart
@@ -1,9 +1,9 @@
-// Openapi Generator last run: : 2026-04-27T11:40:49.753339
+// Openapi Generator last run: : 2026-04-27T13:08:38.512906
 import 'package:openapi_generator_annotations/openapi_generator_annotations.dart';
 
 @Openapi(
   inputSpec: InputSpec(path: '../api-specs/story-service/openapi.yaml'),
   generatorName: Generator.dio,
-  outputDirectory: 'lib/proxy/generated/story_service',
+  outputDirectory: 'packages/openapi',
 )
 class ApiClient {}
