+++ pallas_app/packages/openapi/.gitignore
