+++ pallas_app/packages/openapi/.openapi-generator-ignore
