+++ pallas_app/packages/openapi/.openapi-generator/VERSION
