+++ pallas_app/packages/openapi/doc/Error.md
