+++ pallas_app/packages/openapi/doc/ListStories200Response.md
