+++ pallas_app/packages/openapi/doc/StoriesApi.md
