+++ pallas_app/packages/openapi/doc/Story.md
@@ -1,4 +1,4 @@
-# openapi.model.StoryInput
+# openapi.model.Story
 
 ## Load the model package
 ```dart
@@ -8,7 +8,8 @@ import 'package:openapi/api.dart';
 ## Properties
 Name | Type | Description | Notes
 ------------ | ------------- | ------------- | -------------
-**story** | **String** | The story content | 
+**id** | **String** | Unique identifier for the story | 
+**story** | **String** | The story content as a Quill Delta JSON string (JSON-serialised array of Delta ops). | 
 
 [[Back to Model list]](../README.md#documentation-for-models) [[Back to API list]](../README.md#documentation-for-api-endpoints) [[Back to README]](../README.md)
