+++ pallas_app/packages/openapi/doc/StoryInput.md
@@ -1,4 +1,4 @@
-# openapi.model.Story
+# openapi.model.StoryInput
 
 ## Load the model package
 ```dart
@@ -8,8 +8,7 @@ import 'package:openapi/api.dart';
 ## Properties
 Name | Type | Description | Notes
 ------------ | ------------- | ------------- | -------------
-**id** | **String** | Unique identifier for the story | 
-**story** | **String** | The story content | 
+**story** | **String** | The story content as a Quill Delta JSON string (JSON-serialised array of Delta ops). | 
 
 [[Back to Model list]](../README.md#documentation-for-models) [[Back to API list]](../README.md#documentation-for-api-endpoints) [[Back to README]](../README.md)
