+++ pallas_app/packages/openapi/lib/openapi.dart
