+++ pallas_app/packages/openapi/lib/src/api.dart
