+++ pallas_app/packages/openapi/lib/src/api/stories_api.dart
