+++ pallas_app/packages/openapi/lib/src/api_util.dart
