+++ pallas_app/packages/openapi/lib/src/auth/api_key_auth.dart
