+++ pallas_app/packages/openapi/lib/src/auth/auth.dart
