+++ pallas_app/packages/openapi/lib/src/auth/basic_auth.dart
