+++ pallas_app/packages/openapi/lib/src/auth/bearer_auth.dart
