+++ pallas_app/packages/openapi/lib/src/auth/oauth.dart
