+++ pallas_app/packages/openapi/lib/src/model/error.dart
