+++ pallas_app/packages/openapi/lib/src/model/list_stories200_response.dart
