+++ pallas_app/packages/openapi/lib/src/serializers.dart
