+++ pallas_app/packages/openapi/pubspec.yaml
@@ -1,16 +1,18 @@
----
 name: openapi
 version: 1.0.0
 description: OpenAPI API client
 homepage: homepage
+
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=2.15.0 <4.0.0'
+
 dependencies:
   dio: '^5.2.0'
   one_of: '>=1.5.0 <2.0.0'
   one_of_serializer: '>=1.5.0 <2.0.0'
   built_value: '>=8.4.0 <9.0.0'
   built_collection: '>=5.1.1 <6.0.0'
+
 dev_dependencies:
   built_value_generator: '>=8.4.0 <9.0.0'
   build_runner: any
