+++ pallas_app/packages/openapi/test/error_test.dart
